# 391 more smt solver functionality