--- a/test/rpc-integration/test-equalK-conditions/state-a.execute
+++ b/test/rpc-integration/test-equalK-conditions/state-a.execute
@@ -1,0 +1,1 @@
+state-a.json