--- a/test/rpc-integration/test-equalK-conditions/state-c.execute
+++ b/test/rpc-integration/test-equalK-conditions/state-c.execute
@@ -1,0 +1,1 @@
+state-c.json