… library/Booster/SMT/LowLevelCodec.hs
… library/Booster/SMT/Translate.hs
… scripts/integration-tests.sh
… test/rpc-integration/resources/equalK-conditions.k
… test/rpc-integration/resources/equalK-conditions.kore
… test/rpc-integration/test-diamond/response-infeasible-branching.booster-dev
… test/rpc-integration/test-equalK-conditions/response-a.json
… test/rpc-integration/test-equalK-conditions/response-c.json
… test/rpc-integration/test-equalK-conditions/state-a.json
… test/rpc-integration/test-equalK-conditions/state-c.json
… test/rpc-integration/test-substitutions/response-concrete-substitution.booster-dev
… test/rpc-integration/test-substitutions/response-concrete-substitution.json
… test/rpc-integration/test-substitutions/response-symbolic-substitution.booster-dev
… test/rpc-integration/test-vacuous/response-vacuous-at-branch.json
… test/rpc-integration/test-vacuous/response-vacuous-var-at-branch.json
… tools/booster/Server.hs
… tools/kore-rpc-dev/Server.hs
… unit-tests/Test/Booster/Pattern/ApplyEquations.hs
… unit-tests/Test/Booster/SMT/LowLevel.hs