Fixes #391  
Fixes #390 

- Comments inserted into transcript to explain what things are
- `ensures` clauses are now checked with the SMT solver (changes integration test expectations)
- a test using `==K` (currently fails on `booster-dev`, `kore-rpc`  appears to use internal `==K` logic instead of `z3` to solve it)
- supplying SMT options to `kore-rpc` via the `SMTOption` type in the new code
- separate kill switch `--no-booster-smt` to disable the SMT solver only in booster code but keep the one in `kore-rpc`
- ~SMT support for equation evaluation~ reverted