@jberthold 2023-11-30
Measurements: (`PYTEST_PARALLEL=3` for a change)
### KEVM

| Test                               | 391-more-SMT-solver-functionality time | master-f2ae599 time | (391-more-SMT-solver-functionality/master-f2ae599) time 
|------------------------------------|----------------------------------------|---------------------|---------------------------------------------------------
| erc20/ds/transfer-success-2-spec.k | 68.16                                  | 74.98               | 0.9090424113096824                                      
| erc20/ds/transfer-success-1-spec.k | 82.09                                  | 76.54               | 1.0725111053044158                                      
| TOTAL                              | 150.25                                 | 151.52              | 0.991618268215417

### Kontrol
| Test                           | 391-more-SMT-solver-functionality time | master-f2ae599 time | (391-more-SMT-solver-functionality/master-f2ae599) time 
|--------------------------------|----------------------------------------|---------------------|---------------------------------------------------------
| AssertTest.test_assert_false() | 55.81                                  | 57.92               | 0.9635704419889503                                      
| TOTAL                          | 55.81                                  | 57.92               | 0.9635704419889503    

(surprisingly regular)