@jberthold 2023-12-01
New measurements (`PYTEST_PARALLEL=3`) after adding SMT support for simplifications

### Kontrol: no big effect

| Test                                                     | 391-more-SMT-solver-functionality time | master-f2ae599 time | (391-more-SMT-solver-functionality/master-f2ae599) time 
|----------------------------------------------------------|----------------------------------------|---------------------|---------------------------------------------------------
| AssertTest.test_assert_true()                            | 57.18                                  | 63.18               | 0.9050332383665717                                      
| ExpectRevertTest.testFail_expectRevert_multipleReverts() | 100.33                                 | 107.45              | 0.9337366216845043                                      
| StoreTest.testGasStoreColdVM()                           | 52.55                                  | 55.47               | 0.9473589327564449                                      
| ExpectRevertTest.testFail_expectRevert_false()           | 85.89                                  | 90.08               | 0.9534857904085258                                      
| ExpectRevertTest.testFail_ExpectRevert_failAndSuccess()  | 88.3                                   | 92.3                | 0.9566630552546046                                      
| ExpectRevertTest.test_expectRevert_message()             | 91.6                                   | 95.42               | 0.9599664640536575                                      
| BMCLoopsTest.test_countdown_symbolic(uint256)            | 109.65                                 | 113.68              | 0.9645496129486277                                      
| BlockParamsTest.testChainId(uint256)                     | 45.46                                  | 43.43               | 1.0467418834906748                                      
| BlockParamsTest.testFee(uint256)                         | 44.82                                  | 42.73               | 1.0489117715890477                                      
| SymbolicStorageTest.testFail_SymbolicStorage(uint256)    | 117.73                                 | 111.94              | 1.0517241379310345                                      
| AssumeTest.testFail_assume_true(uint256,uint256)         | 66.84                                  | 63.13               | 1.0587676223665452                                      
| AssumeTest.test_assume_true(uint256,uint256)             | 48.19                                  | 45.34               | 1.0628584031760033                                      
| AssumeTest.testFail_assume_false(uint256,uint256)        | 51.42                                  | 48.22               | 1.0663625051845709                                      
| StoreTest.testGasStoreWarmUp()                           | 67.84                                  | 63.24               | 1.0727387729285263                                      
| ArithmeticTest.test_max2(uint256,uint256)                | 71.2                                   | 65.81               | 1.0819024464367117                                      
| TOTAL                                                    | 1099.00                     | 1101.42             | 0.9978028363385449 

### kevm: some slowdown (will repeat the measurement and edit)

| Test                                     | 391-more-SMT-solver-functionality time | master-f2ae599 time | (391-more-SMT-solver-functionality/master-f2ae599) time 
|------------------------------------------|----------------------------------------|---------------------|---------------------------------------------------------
| mcd/flipper-addu48u48-fail-rough-spec.k  | 51.8                                   | 70.64               | 0.733295583238958                                       
| mcd/end-subuu-pass-spec.k                | 55.96                                  | 73.37               | 0.7627095543137522                                      
| erc20/ds/transfer-failure-2-b-spec.k     | 49.57                                  | 63.79               | 0.7770810471860793                                      
| erc20/ds/allowance-spec.k                | 55.69                                  | 65.49               | 0.8503588334096809                                      
| mcd/cat-exhaustiveness-spec.k            | 108.82                                 | 126.64              | 0.8592861655085281                                      
| mcd/dsvalue-read-pass-summarize-spec.k   | 116.23                                 | 134.39              | 0.8648708981323017                                      
| bihu/functional-spec.k                   | 105.61                                 | 111.66              | 0.945817660755866                                       
| mcd/dsvalue-read-pass-spec.k             | 74.16                                  | 77.49               | 0.9570267131242741                                      
| mcd/flopper-file-pass-rough-spec.k       | 477.98                                 | 495.39              | 0.9648559720624156                                      
| erc20/ds/transferFrom-failure-1-d-spec.k | 67.54                                  | 65.25               | 1.035095785440613                                       
| benchmarks/staticloop00-a0lt10-spec.k    | 47.66                                  | 45.84               | 1.0397033158813263                                      
| erc20/hkg/transfer-failure-1-spec.k      | 59.5                                   | 57.06               | 1.0427620049071153                                      
| benchmarks/ecrecover00-sigvalid-spec.k   | 98.2                                   | 92.81               | 1.0580756384010344                                      
| mcd/flipper-bids-pass-rough-spec.k       | 116.31                                 | 109.47              | 1.0624828720197315                                      
| erc20/ds/transfer-success-2-spec.k       | 89.92                                  | 84.62               | 1.0626329472937839                                      
| erc20/hkg/transfer-failure-2-spec.k      | 58.1                                   | 54.65               | 1.0631290027447393                                      
| erc20/ds/transfer-failure-2-a-spec.k     | 70.42                                  | 66.0                | 1.066969696969697                                       
| erc20/ds/balanceOf-spec.k                | 50.61                                  | 47.42               | 1.0672711935892027                                      
| erc20/ds/transferFrom-failure-1-c-spec.k | 106.49                                 | 99.72               | 1.0678900922583232                                      
| erc20/ds/transferFrom-failure-2-a-spec.k | 78.06                                  | 73.01               | 1.0691686070401314                                      
| erc20/hkg/transfer-success-2-spec.k      | 55.99                                  | 52.23               | 1.0719892781926097                                      
| erc20/ds/transfer-failure-1-a-spec.k     | 73.68                                  | 68.5                | 1.0756204379562044                                      
| benchmarks/ecrecover00-siginvalid-spec.k | 88.68                                  | 82.41               | 1.0760829996359667                                      
| erc20/hkg/transferFrom-failure-1-spec.k  | 82.67                                  | 76.44               | 1.0815018315018317                                      
| erc20/hkg/transfer-success-1-spec.k      | 58.34                                  | 53.9                | 1.0823747680890539                                      
| erc20/hkg/balanceOf-spec.k               | 48.41                                  | 44.63               | 1.0846963925610575                                      
| erc20/ds/transferFrom-failure-2-b-spec.k | 95.46                                  | 87.84               | 1.0867486338797814                                      
| erc20/hkg/approve-spec.k                 | 52.74                                  | 48.51               | 1.0871985157699444                                      
| mcd/vat-deny-diff-fail-rough-spec.k      | 95.06                                  | 87.04               | 1.092141544117647                                       
| erc20/hkg/transferFrom-success-2-spec.k  | 127.91                                 | 116.79              | 1.09521363130405                                        
| examples/erc20-spec.md                   | 228.66                                 | 207.57              | 1.1016042780748663                                      
| erc20/hkg/transferFrom-failure-2-spec.k  | 76.18                                  | 69.04               | 1.1034183082271147                                      
| mcd/vat-move-diff-rough-spec.k           | 136.76                                 | 123.73              | 1.1053099490826799                                      
| benchmarks/bytes00-spec.k                | 81.36                                  | 73.54               | 1.1063366875169975                                      
| erc20/ds/transfer-failure-1-b-spec.k     | 97.58                                  | 87.71               | 1.1125299281723864                                      
| erc20/ds/transfer-success-1-spec.k       | 95.05                                  | 85.18               | 1.1158722704860295                                      
| benchmarks/encode-keccak00-spec.k        | 55.69                                  | 49.4                | 1.1273279352226722                                      
| erc20/ds/approve-success-spec.k          | 64.64                                  | 57.25               | 1.1290829694323143                                      
| benchmarks/keccak00-spec.k               | 54.42                                  | 47.42               | 1.1476170392239562                                      
| erc20/ds/transferFrom-failure-1-b-spec.k | 148.13                                 | 128.14              | 1.1560012486343063                                      
| erc20/hkg/transferFrom-success-1-spec.k  | 78.27                                  | 67.1                | 1.166467958271237                                       
| erc20/ds/transferFrom-success-1-spec.k   | 110.31                                 | 92.25               | 1.1957723577235773                                      
| erc20/ds/transferFrom-success-2-spec.k   | 98.91                                  | 80.56               | 1.227780536246276                                       
| erc20/ds/transferFrom-failure-1-a-spec.k | 109.64                                 | 86.36               | 1.269569245020843                                       
| mcd/dsvalue-peek-pass-rough-spec.k       | 76.83                                  | 60.48               | 1.2703373015873016                                      
| mcd/flopper-cage-pass-spec.k             | 75.74                                  | 59.44               | 1.2742261103633916                                      
| erc20/hkg/allowance-spec.k               | 66.08                                  | 49.06               | 1.3469221361598043                                      
| erc20/ds/approve-failure-spec.k          | 67.96                                  | 47.87               | 1.4196782953833298                                      
| bihu/forwardToHotWallet-failure-1-spec.k | 81.87                                  | 46.87               | 1.7467463196074249                                      
| TOTAL                                    | 4521.65                     | 4251.97             | 1.0634247184246357          

In the worst case we can revert the SMT support for equation evaluation for now.