@jberthold 2023-12-01
Re-testing without parallelism still shows a slow-down so I reverted that change.

| Test                                     | 391-more-SMT-solver-functionality time | master-f2ae599 time | (391-more-SMT-solver-functionality/master-f2ae599) time 
|------------------------------------------|----------------------------------------|---------------------|---------------------------------------------------------
| mcd/vat-slip-pass-rough-spec.k           | 128.2                                  | 123.66              | 1.0367135694646612                                      
| erc20/hkg/transfer-failure-1-spec.k      | 54.35                                  | 52.1                | 1.0431861804222649                                      
| erc20/hkg/transfer-failure-2-spec.k      | 52.81                                  | 50.31               | 1.0496919101570263                                      
| benchmarks/ecrecover00-sigvalid-spec.k   | 87.98                                  | 83.49               | 1.053778895676129                                       
| mcd/flipper-bids-pass-rough-spec.k       | 102.97                                 | 97.71               | 1.053832770443148                                       
| benchmarks/ecrecover00-siginvalid-spec.k | 80.31                                  | 75.7                | 1.060898282694848                                       
| erc20/ds/transferFrom-failure-1-a-spec.k | 80.04                                  | 75.21               | 1.0642201834862388                                      
| erc20/ds/transfer-failure-1-a-spec.k     | 67.23                                  | 62.89               | 1.0690093814596915                                      
| erc20/ds/balanceOf-spec.k                | 46.49                                  | 43.47               | 1.0694731999079825                                      
| erc20/ds/transferFrom-failure-2-a-spec.k | 70.59                                  | 65.95               | 1.0703563305534496                                      
| erc20/hkg/balanceOf-spec.k               | 44.2                                   | 41.29               | 1.0704771131024462                                      
| erc20/ds/transferFrom-failure-1-c-spec.k | 96.21                                  | 89.78               | 1.0716195143684561                                      
| erc20/hkg/transfer-success-1-spec.k      | 52.8                                   | 49.0                | 1.0775510204081633                                      
| erc20/hkg/transfer-success-2-spec.k      | 50.8                                   | 47.05               | 1.079702444208289                                       
| erc20/hkg/approve-spec.k                 | 48.04                                  | 44.47               | 1.0802788396671914                                      
| erc20/ds/transfer-failure-2-a-spec.k     | 64.16                                  | 59.38               | 1.0804984843381609                                      
| erc20/hkg/transferFrom-failure-1-spec.k  | 75.7                                   | 69.88               | 1.0832856325128792                                      
| erc20/ds/transferFrom-failure-2-b-spec.k | 86.17                                  | 79.33               | 1.0862221101726963                                      
| erc20/hkg/transferFrom-success-2-spec.k  | 116.25                                 | 107.0               | 1.0864485981308412                                      
| mcd/vat-deny-diff-fail-rough-spec.k      | 87.34                                  | 79.81               | 1.0943490790627741                                      
| erc20/hkg/transferFrom-failure-2-spec.k  | 69.57                                  | 63.55               | 1.0947285601888277                                      
| mcd/vat-move-diff-rough-spec.k           | 124.18                                 | 113.13              | 1.0976752408733317                                      
| erc20/ds/transfer-failure-1-b-spec.k     | 88.23                                  | 80.18               | 1.100399102020454                                       
| erc20/ds/allowance-spec.k                | 50.72                                  | 45.88               | 1.105492589363557                                       
| benchmarks/encode-keccak00-spec.k        | 50.08                                  | 45.18               | 1.1084550686144312                                      
| erc20/ds/transfer-success-2-spec.k       | 66.24                                  | 59.04               | 1.121951219512195                                       
| erc20/ds/approve-success-spec.k          | 58.61                                  | 52.22               | 1.12236690923018                                        
| erc20/ds/transfer-success-1-spec.k       | 69.88                                  | 62.18               | 1.1238340302348022                                      
| erc20/ds/transferFrom-failure-1-b-spec.k | 131.49                                 | 115.96              | 1.1339254915488102                                      
| benchmarks/keccak00-spec.k               | 48.91                                  | 42.63               | 1.147314098053014                                       
| erc20/hkg/transferFrom-success-1-spec.k  | 70.18                                  | 60.8                | 1.1542763157894738                                      
| benchmarks/bytes00-spec.k                | 57.98                                  | 49.89               | 1.162156744838645                                       
| erc20/ds/transferFrom-success-1-spec.k   | 98.7                                   | 82.39               | 1.1979609175870858                                      
| erc20/ds/transferFrom-success-2-spec.k   | 88.64                                  | 72.55               | 1.2217780840799448                                      
| erc20/hkg/allowance-spec.k               | 59.15                                  | 45.42               | 1.302289740202554                                       
| TOTAL                                    | 2625.2                                 | 2388.48 | 1.0991090568060018